+++ services/jwt-manager/jwt-manager.js
